fixed issue when input is int

+++ autoregressive/datasets/fourier_dataset.py
@@ -126,7 +126,7 @@ class FSeriesDataset(sd.SeriesDataset):
 
 def add_period_conditioning(sm: sd.SeriesMeta, max_period: int = 5) -> sd.SeriesMeta:
     series, meta = sm
-    p = torch.round(meta["period"]).long()
+    p = torch.round(meta["period"].float()).long()
     p = F.one_hot(p, num_classes=max_period + 1).permute(1, 0)
     series["c"] = p.float()
     return series, meta
